+++ tools/vf-core/fractal.js
@@ -1,30 +1,32 @@
+const renderMarkdown = require("marked");
+
 module.exports = {
 
   // mode: 'build' or 'server'
   initialize: function(mode, callback) {
 
     /* Create a new Fractal instance and export it for use elsewhere if required */
-    const fractal        = module.exports = require('@frctl/fractal').create();
+    const fractal        = module.exports = require("@frctl/fractal").create();
     const logger         = fractal.cli.console;
-    var vfName           = global.vfName || 'Visual Framework component library';
+    var vfName           = global.vfName || "Visual Framework component library";
     const projectTitle   = vfName;
-    const path           = require('path');
+    const path           = require("path");
 
     /* Set the title of the project */
-    fractal.set('project.title', projectTitle);
+    fractal.set("project.title", projectTitle);
 
     /* Tell Fractal where the components will live */
-    var vfComponentPath = global.vfComponentPath || __dirname + '/components';
+    var vfComponentPath = global.vfComponentPath || __dirname + "/components";
     // Possible needed fix to resolve path
     // vfComponentPath = path.resolve('.',vfComponentPath);
 
-    fractal.components.set('path', vfComponentPath);
+    fractal.components.set("path", vfComponentPath);
 
     /* Tell Fractal where the documentation pages will live */
-    var vfDocsPath = global.vfDocsPath || __dirname + '/docs';
-    fractal.docs.set('path', vfDocsPath);
+    var vfDocsPath = global.vfDocsPath || __dirname + "/docs";
+    fractal.docs.set("path", vfDocsPath);
 
-    const nunj = require('@frctl/nunjucks')({
+    const nunj = require("@frctl/nunjucks")({
       env: {
         lstripBlocks: true,
         trimBlocks: true,
@@ -35,8 +37,7 @@ module.exports = {
       filters: {
         // {{ "## Parse me" | marked }}
         marked: function(string) {
-          const renderMarkdown = require('marked');
-          return renderMarkdown(string);
+          return renderMarkdown.parse(string);
         },
         // A filter and non-async version of frctl's context extension from
         // https://github.com/frctl/nunjucks/blob/develop/src/extensions/context.js
@@ -55,61 +56,61 @@ module.exports = {
         hextorgb: module.exports = function(text) {
           function hexToRGB(hex) {
             var r = parseInt(hex.slice(1, 3), 16),
-                g = parseInt(hex.slice(3, 5), 16),
-                b = parseInt(hex.slice(5, 7), 16);
+              g = parseInt(hex.slice(3, 5), 16),
+              b = parseInt(hex.slice(5, 7), 16);
             return "rgb(" + r + ", " + g + ", " + b + ")";
           }
 
           var hex = new String(text);
 
           return hexToRGB(hex);
-          }
+        }
 
       },
       // globals: {
       //   // global-name: global-val
       // },
       extensions: {
-        codeblock: require('@visual-framework/vf-frctl-extensions/codeblock.js')(fractal),
-        spaceless: require('@visual-framework/vf-frctl-extensions/spaceless.js')(fractal),
-        markdown:  require('@visual-framework/vf-frctl-extensions/markdown.js')(fractal)
+        codeblock: require("@visual-framework/vf-frctl-extensions/codeblock.js")(fractal),
+        spaceless: require("@visual-framework/vf-frctl-extensions/spaceless.js")(fractal),
+        markdown:  require("@visual-framework/vf-frctl-extensions/markdown.js")(fractal)
       }
     });
 
-    fractal.components.set('ext', '.njk'); // look for files with a .nunj file extension
+    fractal.components.set("ext", ".njk"); // look for files with a .nunj file extension
     fractal.components.engine(nunj); /* set as the default template engine for components */
-    fractal.docs.set('ext', '.njk'); // look for files with a .njk file extension
+    fractal.docs.set("ext", ".njk"); // look for files with a .njk file extension
     fractal.docs.engine(nunj); /* you can also use the same instance for documentation, if you like! */
 
     /* configure components */
-    fractal.components.set('default.status', 'alpha');
-    fractal.components.set('default.preview', `@preview`);
+    fractal.components.set("default.status", "alpha");
+    fractal.components.set("default.preview", "@preview");
 
     /* build destination */
-    var vfBuilderPath = global.vfBuilderPath || __dirname + '/build';
-    fractal.web.set('builder.dest', vfBuilderPath);
-    fractal.set('components.resources.assets.match',  ['**/*.njk', '**/*.config.yml', '**/*.scss', '**/CHANGELOG.md', '**/*.js','**/*.css', '!**/*.precompiled.js', '!**/package.variables.scss', '!**/index.scss'])
+    var vfBuilderPath = global.vfBuilderPath || __dirname + "/build";
+    fractal.web.set("builder.dest", vfBuilderPath);
+    fractal.set("components.resources.assets.match",  ["**/*.njk", "**/*.config.yml", "**/*.scss", "**/CHANGELOG.md", "**/*.js","**/*.css", "!**/*.precompiled.js", "!**/package.variables.scss", "!**/index.scss"]);
 
     /* configure web */
-    var vfStaticPath = global.vfStaticPath || __dirname + '/temp/build-files';
-    fractal.web.set('static.path', vfStaticPath);
-    fractal.web.set('server.sync', true);
+    var vfStaticPath = global.vfStaticPath || __dirname + "/temp/build-files";
+    fractal.web.set("static.path", vfStaticPath);
+    fractal.web.set("server.sync", true);
     var vfOpenBrowser = typeof global.vfOpenBrowser === "undefined" ? true : global.vfOpenBrowser;
-    fractal.web.set('server.syncOptions', {
+    fractal.web.set("server.syncOptions", {
       watchOptions: {
         // @todo: this should use vfComponentPath
-        ignored: path.join(__dirname, '../../components/**/*.scss'),
+        ignored: path.join(__dirname, "../../components/**/*.scss"),
       },
       open: vfOpenBrowser,
-      browser: 'default',
+      browser: "default",
       sync: true
     });
 
-    var vfThemePath = global.vfThemePath || '@frctl/mandelbrot';
+    var vfThemePath = global.vfThemePath || "@frctl/mandelbrot";
     const vfTheme = require(vfThemePath);
     const vfThemeConfig = vfTheme({}, fractal);
 
-    fractal.components.set('statuses', {
+    fractal.components.set("statuses", {
       /* status definitions here */
       alpha: {
         label: "alpha",
@@ -136,8 +137,8 @@ module.exports = {
 
     fractal.web.theme(vfThemeConfig);
 
-    if (mode == 'server') {
-      fractal.set('project.environment.local', 'true');
+    if (mode == "server") {
+      fractal.set("project.environment.local", "true");
       const fractalServer = fractal.web.server({
         sync: true
       });
@@ -150,27 +151,27 @@ module.exports = {
     }
 
     // If you want to build static html files
-    if (mode == 'build') {
-      fractal.set('project.environment.production', 'true');
+    if (mode == "build") {
+      fractal.set("project.environment.production", "true");
       const builder = fractal.web.builder();
-      builder.on('progress', (completed, total) =>
-        logger.update(`Exported ${completed} of ${total} items`, 'info')
+      builder.on("progress", (completed, total) =>
+        logger.update(`Exported ${completed} of ${total} items`, "info")
       );
-      builder.on('error', err => logger.error(err.message));
+      builder.on("error", err => logger.error(err.message));
       return builder.build().then(() => {
-        logger.success('Fractal build completed!');
+        logger.success("Fractal build completed!");
 
         callback(fractal);
       });
     }
 
     // To build the fractal object in memory
-    if (mode == 'dataobject') {
-      fractal.set('project.environment.production', 'true');
+    if (mode == "dataobject") {
+      fractal.set("project.environment.production", "true");
       const server = fractal.web.server();
 
       server.start().then(function(){
-        console.log('The Fractal component data has been generated.');
+        console.log("The Fractal component data has been generated.");
         callback(fractal);
         server.stop();
         fractal.unwatch(); // exit fractal
@@ -178,4 +179,4 @@ module.exports = {
 
     }
   }
-}
+};